feat: 회원가입 상태 공유구현
