+++ src/routes/index.jsx
@@ -10,6 +10,7 @@ import KakaoCheck from './pages/KakaoCheck';
 import Error from './pages/Error';
 import PasswordFind from './pages/PasswordFind';
 import SignUp from './pages/SignUp/SignUp';
+import SignUpHome from './pages/SignUp/SignUpHome';
 import GetLocation from './pages/GetLocation';
 const router = createBrowserRouter([
   {
@@ -54,10 +55,11 @@ const router = createBrowserRouter([
   },
   {
     path: '/sign-up',
+    Component: SignUp,
     children: [
       {
         index: true,
-        Component: SignUp,
+        Component: SignUpHome,
       },
       {
         path: 'get-location',
